proxy_set_header_map_pairs doesn't add headers to the request/response correctly
From the docs:
```
Sets request headers with the given header pairs. For each header key
value pair, it acts the same way as replaceRequestHeader.
```
Currently proxy_set_header_map_pairs is not exporting the pairs to be set and
setHeaderMapPairs is deleting all the existing headers. This pr fixes export of
pairs and replacing of headers in the map

Signed-off-by: Mohit Agarwal <mohit@traceable.ai>