Add support for all headers statuses returned by WASM.
This requires the changes made to proxy-wasm-cpp-sdk
in https://github.com/proxy-wasm/proxy-wasm-cpp-sdk/pull/19.

With this change, a WASM filter can return, say,
StopAllIterationAndWatermark. This is useful when a WASM filter
might want to call an external service and use the result to
change an HTTP header before continuing processing the body.

Signed-off-by: Gregory Brail <gregbrail@google.com>